Read node ids at their full width in the extensionality layer
GetNodeNum() now returns uint64_t rather than unsigned, so the three
places here that received it into a 32-bit type reintroduced, inside
this layer, exactly the truncation that widening the accessor removed.

The reified equality cache is the one that can change an answer. Its
key is a node-number pair, so two distinct operand pairs whose ids
agree modulo 2^32 share an entry, and the second lookup hands back the
literal reifying the *first* pair's equality. That literal goes on to
stand for the wrong atom in a lemma clause, which is a silent wrong
answer rather than a crash.

The other two only cost determinism. Premise atoms are ordered by
operand node number, and the per-source adjacency used to walk equality
edges is sorted on the far endpoint's; truncated, either can tie two
entries that are not equal, leaving their relative order to the sort.
The premise ordering is what makes a lemma's text reproducible, and the
edge ordering feeds the shortest-path property the path guards rely on.

Reaching any of this needs 2^31 nodes on one thread, which a single
command-line solve will not do but a long-lived library user can.

+++ include/stp/Extensionality/ExtensionalityContext.h
@@ -437,7 +437,7 @@ private:
                       ToSATBase* tosat);
 
   // reified equality cache, scoped to the current SAT instance
-  std::map<std::pair<unsigned, unsigned>, int> eqLitCache;
+  std::map<std::pair<uint64_t, uint64_t>, int> eqLitCache;
 
   // last consistent observations for model export
   std::map<ASTNode, std::vector<std::pair<ASTNode, ASTNode>>> lastObserved;
